Reorganise project structure to highlight the part that should be extracted as a library
