+++ src/app/components/Project.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Alert, Avatar, Badge, Card, Heading, Pane, Paragraph, majorScale } from 'evergreen-ui';
 
 import { onlyWhen } from '../utils/javascriptUtils';
-import { SemanticComponentBuilder } from '../services/SemanticComponentBuilder';
+import { SemanticComponentBuilder } from '../../library/services/SemanticComponentBuilder';
 import Semantics from '../utils/semantics';
 
 const Project = ({title, isPublic, lastUpdate, collaborators}) =>
