+++ src/app/utils/javascriptUtils.js
