--- src/components/Actions.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import { Pane, Heading, majorScale } from 'evergreen-ui';
-
-const Actions = ({semanticData}) => {
-
-  return <Pane width="100%" marginBottom={majorScale(4)}>
-    <Heading size={500}>Actions</Heading>
-
-  </Pane>
-}
-
-export default Actions;
