--- src/hooks/documentationHooks.js
@@ -1,9 +1,0 @@
-import { useMemo } from 'react';
-
-export const useRequestBodySchema = (apiDocumentation, operation) => useMemo(
-  () => operation && operation.requestBody
-    ? apiDocumentation.requestBodySchema(operation)
-    : undefined,
-  [apiDocumentation, operation]
-);
-
